server/uhf-bridge/requirements.txt
flask>=2.0
flask-cors>=3.0
pyserial>=3.4
# uhfReaderApi requires pyserial==3.4; install separately if needed:
#   pip install uhfReaderApi
# The bridge runs in DEMO mode without it.
